Added add,delete actions

+++ Elective_Business_logic/Models/Instructor.cs
@@ -8,7 +8,7 @@ namespace Elective.Models
 {
     public class Instructor : Account
     {
-        public Instructor(byte id, string firstname, string lastname) : base(id, firstname, lastname)
+        public Instructor(int id, string firstname, string lastname) : base(id, firstname, lastname)
         {
 
         }
